requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
anthropic
requests
ffmpeg-python
python-dotenv
tenacity
pandas
pytest
Pillow
